investment_researcher/pyproject.toml
[project]
name = "investment-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "yfinance>=1.2.0",
    "streamlit>=1.45.0",
]
